fix: use float type

+++ src/pytemplate/domain/models.py
@@ -3,5 +3,5 @@ from dataclasses import dataclass
 
 @dataclass
 class Operands:
-    first_operand: int
+    first_operand: float
     second_operand: int
